RegistationCOntroller extracted from the Authentication controller
