+++ authentication/src/main/java/com/microservices/authentication/controller/AuthenticationController.java
@@ -2,7 +2,6 @@ package com.microservices.authentication.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -12,13 +11,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.microservices.authentication.dto.LoginRequest;
 import com.microservices.authentication.dto.LoginResponse;
-import com.microservices.authentication.dto.OTPVerifyRequest;
-import com.microservices.authentication.dto.RegistrationRequest;
-import com.microservices.authentication.dto.RetailerRegister;
 import com.microservices.authentication.service.LoginService;
-import com.microservices.authentication.service.RegistrationService;
-
-import jakarta.validation.Valid;
 
 @RestController
 @RequestMapping("/auth")
@@ -27,35 +20,9 @@ public class AuthenticationController {
 	@Autowired
 	private LoginService service;
 
-	@Autowired
-	private RegistrationService registrationService;
-
 	@PostMapping("/login")
 	public ResponseEntity<LoginResponse> login(@RequestBody LoginRequest request) {
 		return ResponseEntity.ok(service.login(request.getEmail(), request.getPassword()));
-	}
-
-	@GetMapping("/register/verify-email")
-	public ResponseEntity<String> verifyEmail(@RequestParam String email) {
-		String message = registrationService.verifyEmail(email);
-		return ResponseEntity.ok(message);
-	}
-
-	@PostMapping("/register/verify-email")
-	public ResponseEntity<String> verifyEmail(@RequestBody OTPVerifyRequest otpVerifyRequest) {
-		String message = registrationService.verifyEmail(otpVerifyRequest.getEmail(), otpVerifyRequest.getOtp());
-		return ResponseEntity.ok(message);
-	}
-
-	@PostMapping("/register")
-	public String registerCustomer(@RequestBody @Valid RegistrationRequest request) {
-		return registrationService.register(request);
-	}
-
-	@PostMapping("/register-retailer")
-	public ResponseEntity<String> registerUser(@RequestBody @Valid RetailerRegister request) {
-		String msg = registrationService.registerRetailer(request);
-		return ResponseEntity.ok(msg);
 	}
 
 	@PostMapping("/forgotpassword")
